for final

+++ practice/w6-practice/w6/w6/Car.h
@@ -9,7 +9,7 @@ namespace sdds
 	{
 		std::string m_model;
 		std::string m_condition;
-		int m_topSpeed;
+		double m_topSpeed;
 	protected:
 		std::string& trim(std::string& token);
 	public:
